WMLX——dg-1688-skill/requirements.txt
xlsxwriter>=3.0.0
Pillow>=9.0.0
requests>=2.28.0
